# doc/openapi.yaml
openapi: 3.0.3
info:
  title: Projeto do Filmes DS2M - Matheus Aguiar
  description: API destinada para controlar Filmes
  version: 1.0.4.26
servers:
  - url: http://localhost:8080
paths:
...
  /v1/senai/locadora/filme:
    get:
      tags:
        - FILME
      summary: Listar filmes
      responses:
        '200':
          description: Lista de filmes

    post:
      tags:
        - FILME
      summary: Inserir filme
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
      responses:
        '201':
          description: Filme criado
  /v1/senai/locadora/filme/{id}:
    get:
      tags:
        - FILME
      summary: Buscar filme por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Filme encontrado

    put:
      tags:
        - FILME
      summary: Atualizar filme
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
      responses:
        '200':
          description: Filme atualizado

    delete:
      tags:
        - FILME
      summary: Excluir filme
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Filme excluído

# =========================================
# CLASSIFICACAO
# =========================================
...
  /v1/senai/locadora/genero:
    get:
      tags:
        - GENERO
      summary: Listar gêneros
      responses:
        '200':
          description: Lista de gêneros

    post:
      tags:
        - GENERO
      summary: Inserir gênero
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Genero'
      responses:
        '201':
          description: Gênero criado
  /v1/senai/locadora/genero/{id}:
    get:
      tags:
        - GENERO
      summary: Buscar gênero por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Gênero encontrado

    put:
      tags:
        - GENERO
      summary: Atualizar gênero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Genero'
      responses:
        '200':
          description: Gênero atualizado

    delete:
      tags:
        - GENERO
      summary: Excluir gênero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Gênero excluído

# =========================================
# IDIOMA
# =========================================
...
components:
  schemas:
    Filme:
      type: object
      properties:
        nome:
          type: string
        sinopse:
          type: string
        data_lancamento:
          type: string
        duracao:
          type: string
        valor:
          type: number
        avaliacao:
          type: number
        capa:
          type: string
...
    Genero:
      type: object
      properties:
        tipo:
          type: string
        descricao:
          type: string
...